# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Good Food Tracker API",
    "description": "The project aims to allow the users to take pictures and/or leave notes, ratings, comments about restaurants they visit\r\n\tin order to be able to reference it later when they try to pick were they want to go eat out or order from.",
    "contact": {
      "name": "Igor Ilić",
      "email": "github@igorilic.net"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8098",
      "description": "Development API base url"
    },
    {
      "url": "https://gft.igorilic.net/api",
      "description": "Production API base url"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login endpoint",
        "description": "Endpoint used to authenticate user and obtain JWT tokens",
        "operationId": "f37d342cb00753efb2efa7875d4618cf",
        "requestBody": {
          "description": "Login parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/UserEntity"
                    },
                    "tokens": {
                      "properties": {
                        "access_token": {
                          "type": "string"
                        },
                        "refresh_token": {
                          "type": "string"
                        }
                      },
                      "type": "object"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields in the request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "404": {
            "description": "User not found (invalid username/password)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "423": {
            "description": "Account not active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register endpoint",
        "description": "Endpoint used for registering new accounts",
        "operationId": "42e7f1d156b5d98a81e2b94e9a7af221",
        "requestBody": {
          "description": "Register parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "password_again": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "$ref": "#/components/schemas/UserEntity"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields in the request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "409": {
            "description": "Username or Email taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "500": {
            "description": "Registration failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify account endpoint",
        "description": "Endpoint used for verifying  user accounts after registration",
        "operationId": "7b08c8dfe317bea9f1c37a266e9f3ee0",
        "requestBody": {
          "description": "Parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "activationKey": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_with_message_only"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields in the request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "412": {
            "description": "Invalid activation key provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      }
    },
    "/auth/request-password-reset": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request password reset code endpoint",
        "description": "Endpoint used for request a new password reset code",
        "operationId": "bc96bacf353b7d2741615279b0c7e1da",
        "requestBody": {
          "description": "Parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "emailOrUsername": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset code sent to email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_with_message_only"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields in the request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "404": {
            "description": "User account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "500": {
            "description": "Unable to send reset code via email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password endpoint",
        "description": "Endpoint used for resetting a new password",
        "operationId": "cfedc29b38bb0d51eefc48d5c0f22d1e",
        "requestBody": {
          "description": "Parameters",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "passwordResetCode": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  },
                  "newPasswordAgain": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_with_message_only"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields in the request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "404": {
            "description": "User account not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "500": {
            "description": "Unable to send password reset confirmation email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout endpoint",
        "description": "Endpoint used for logging user out and invalidating tokens",
        "operationId": "a658b3822483da1e072726837721c1b1",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Refresh or access token",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logout Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_with_message_only"
                }
              }
            }
          },
          "401": {
            "description": "Missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "498": {
            "description": "Invalid token exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "500": {
            "description": "Invalid or incorrect token provided: InvalidArgumentException | UnexpectedValueException | SignatureInvalidException | BeforeValidException | BeforeValidException | ExpiredException",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh token endpoint",
        "description": "Endpoint used for generating a fresh access token based of refresh token",
        "operationId": "8e0ff1a489c5dd385793c45a6da6f72a",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Refresh token",
            "required": true,
            "schema": {
              "type": "string",
              "additionalProperties": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful generated new access token",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "nullable": true,
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Missing token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "498": {
            "description": "Invalid token exception",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          },
          "500": {
            "description": "Invalid or incorrect token provided: InvalidArgumentException | UnexpectedValueException | SignatureInvalidException | BeforeValidException | BeforeValidException | ExpiredException",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error_response"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserEntity": {
        "description": "UserEntity class",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "image": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "error_response": {
        "properties": {
          "class": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "field": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "response_with_message_only": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}